+++ gateway/dns.go
@@ -1,16 +1,12 @@
 package gateway
 
 import (
-	"context"
 	"fmt"
-	"net"
 	"strings"
-	"time"
 
 	"github.com/libp2p/go-doh-resolver"
 	dns "github.com/miekg/dns"
 	madns "github.com/multiformats/go-multiaddr-dns"
-	"github.com/rs/dnscache"
 )
 
 var defaultResolvers = map[string]string{
@@ -93,59 +89,4 @@ func NewDNSResolver(resolvers map[string]string, dohOpts ...doh.Option) (*madns.
 	}
 
 	return madns.NewResolver(opts...)
-}
-
-// CachedDNS implements [http.Transport.DialContext], allowing to cache DNS
-// requests for a specified amount of time.
-type CachedDNS struct {
-	resolver  *dnscache.Resolver
-	refresher *time.Ticker
-}
-
-func NewCachedDNS(refreshInterval time.Duration) *CachedDNS {
-	cache := &CachedDNS{
-		resolver:  &dnscache.Resolver{},
-		refresher: time.NewTicker(refreshInterval),
-	}
-
-	// Configure DNS cache to not remove stale records to protect gateway from
-	// catastrophic failures like https://github.com/ipfs/bifrost-gateway/issues/34
-	options := dnscache.ResolverRefreshOptions{}
-	options.ClearUnused = false
-	options.PersistOnFailure = true
-
-	// Every refreshInterval we check for updates, but if there is
-	// none, or if domain disappears, we keep the last cached version
-	go func(cdns *CachedDNS) {
-		defer cdns.refresher.Stop()
-		for range cdns.refresher.C {
-			cdns.resolver.RefreshWithOptions(options)
-		}
-	}(cache)
-
-	return cache
-}
-func (cdns *CachedDNS) DialContext(ctx context.Context, network string, addr string) (conn net.Conn, err error) {
-	host, port, err := net.SplitHostPort(addr)
-	if err != nil {
-		return nil, err
-	}
-	ips, err := cdns.resolver.LookupHost(ctx, host)
-	if err != nil {
-		return nil, err
-	}
-	// Try all IPs returned by DNS
-	for _, ip := range ips {
-		var dialer net.Dialer
-		conn, err = dialer.DialContext(ctx, network, net.JoinHostPort(ip, port))
-		if err == nil {
-			break
-		}
-	}
-	return
-}
-
-func (cdns *CachedDNS) Close() error {
-	cdns.refresher.Stop()
-	return nil
 }